app/requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
packaging==24.2
psycopg2==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.7
sniffio==1.3.1
sqlparse==0.5.2
